pyproject.toml
[project]
name = "observo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.12"

dependencies = [
    "celery>=5.5.3",
    "charset-normalizer>=3.4.3",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-health-check>=3.20.0",
    "django-unfold>=0.64.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[blacklist]>=5.5.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "langchain-text-splitters>=0.3.11",
    "pandas>=2.3.2",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pygments>=2.19.2", # For formatting JSON files
    "pytest>=8.4.1",
    "pytz>=2024.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "timezonefinder>=6.5.2",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "ddt>=1.7.2",
    "django-stubs>=5.2.2",
    "factory-boy>=3.3.3",
    "ipython>=9.4.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true  # Formats code in docstrings.

[tool.ruff.lint.isort]
known-first-party = [
    "analytics",
    "observo",
    "utils",
]

[tool.mypy]
mypy_path = "analytics/"
exclude = [
    "migrations/",
    "tests/",
]
plugins = [
    "mypy_django_plugin.main"
]
ignore_missing_imports = true
disallow_untyped_defs = true  # Requires annotations on functions
warn_redundant_casts = true   # Flags unnecessary casts
warn_unused_ignores = true    # Helps clean up # type: ignor

[tool.django-stubs]
django_settings_module = "observo.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "observo.settings"
addopts = "observo/ --cov --cov-report term-missing --cov-fail-under=85"

[tool.coverage.run]
source = [
    "analytics",
    "observo",
    "utils",
]
omit = [
    "*/__init__.py",
    "*/migrations/*",
    "*/tests/*",
]

[tool.coverage.report]
omit = [
    "*/migrations/*",
    "*/tests/*",
]

[tool.codespell]
skip = [
    "*/locale/*"
]
ignore-words-list = [
    "assertIn",
]
